fix: build na integração de senha
